Respond to keyword to start questions
The journey for the user is syncronous and when a user types in
a single keyword it is assumed that they want to begin a new set
of questions. After the question is asked, an entry into the scoring
table is made so that we can track the users progression.