Recieve an answer to a question
Looking for T or F replies indicates an answer. First thing
I did was look for the last entry by the user. If the point
is Nil that meant that the question had not yet been answered
but it had been asked. Then depending on the answer a point
value is inserted by updated that row.